Moves `objectWarnExactlyOneOfChildren` to `objectvalidator.WarnExactlyOneOfChildren`
